+++ Mods ChoGGi/Pipes Less Limits/metadata.lua
@@ -1,0 +1,25 @@
+return PlaceObj("ModDef", {
+	"title", "Pipes Less Limits",
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"saved", 0,
+	"image", "Preview.png",
+	"id", "ChoGGi_PipesLessLimits",
+	"steam_id", "1819506189",
+	"pops_any_uuid", "0fb8383a-6320-4a1f-864e-21b962291424",
+	"author", "ChoGGi",
+	"lua_revision", 245618,
+	"code", {
+		"Code/Script.lua",
+	},
+	"TagGameplay", true,
+	"description", [[Allows more buildings to be built under pipes.
+Subsurface Heater, Tunnel, Landing/Trade Pad
+
+Some won't stick through depending on angle, up to you to make it not stupid:
+Drone Hub, Shuttle Hub, Automatic Metals Extractor, Rare/Metals Refinery, Waste Rock Processor, Polymer Plant, Drone Assembler, Fuel Refinery, MOXIE, Oxygen Tank, Water Extractor, Moisture Vaporator, Fungal Farm, Outside Ranch, Mech Storage
+
+Any I missed?
+]],
+})